doc: update README.md
